# PERF: DataFrame.groupby.nunique